+++ src/init/consented/third-party-tags.spec.ts
@@ -93,7 +93,7 @@ afterEach(() => {
 	document.body.innerHTML = '';
 });
 
-jest.mock('ophan-tracker-js', () => null);
+jest.mock('@guardian/ophan-tracker-js', () => null);
 
 jest.mock('lib/commercial-features', () => ({
 	commercialFeatures: {